app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
click==8.1.8
    # via
    #   -r requirements.in
    #   pip-tools
django==5.1.6
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.4.0
    # via -r requirements.in
packaging==24.2
    # via
    #   -r requirements.in
    #   build
pip-tools==7.4.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
